# docs/core/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LogQSORequest": {
        "description": "Request to log a QSO.",
        "properties": {
          "callsign": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override callsign",
            "title": "Callsign"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional comments",
            "title": "Comments"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override end time",
            "title": "End Time"
          },
          "frequency_hz": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override frequency",
            "title": "Frequency Hz"
          },
          "image_id": {
            "description": "Image ID (public UUID) to create QSO from",
            "format": "uuid",
            "title": "Image Id",
            "type": "string"
          },
          "is_sent": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "description": "True if we initiated contact",
            "title": "Is Sent"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override mode",
            "title": "Mode"
          },
          "report": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override signal report",
            "title": "Report"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override start time",
            "title": "Start Time"
          }
        },
        "required": [
          "image_id"
        ],
        "title": "LogQSORequest",
        "type": "object"
      },
...
      "QSOListResponse": {
        "description": "Paginated QSO list response.",
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "qsos": {
            "items": {
              "$ref": "#/components/schemas/QSOResponse"
            },
            "title": "Qsos",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "qsos",
          "total",
          "limit",
          "offset"
        ],
        "title": "QSOListResponse",
        "type": "object"
      },
      "QSOResponse": {
        "description": "QSO record response.",
        "properties": {
          "callsign": {
            "title": "Callsign",
            "type": "string"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "frequency_hz": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frequency Hz"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Image Ids",
            "type": "array"
          },
          "is_sent": {
            "title": "Is Sent",
            "type": "boolean"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "report": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "callsign",
          "mode",
          "frequency_hz",
          "start_time",
          "end_time",
          "report",
          "comments",
          "is_sent"
        ],
        "title": "QSOResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "SSTV (Slow-Scan Television) encode/decode engine",
    "title": "SSTeVe API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/qso/export": {
      "get": {
        "description": "Export QSOs to ADIF format.\n\nADIF (Amateur Data Interchange Format) is the standard format for\namateur radio logbook data exchange.\n\nArgs:\n    start_date: Optional start date filter\n    end_date: Optional end date filter\n    callsign_filter: Optional callsign substring filter\n    db: Database session\n\nReturns:\n    ADIF format file as plain text",
        "operationId": "export_adif_api_v1_qso_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "callsign_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Callsign Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Adif",
        "tags": [
          "QSO Logging",
          "qso"
        ]
      }
    },
    "/api/v1/qso/list": {
      "get": {
        "description": "List QSOs with filtering and pagination.\n\nArgs:\n    limit: Maximum number of QSOs to return\n    offset: Offset for pagination\n    callsign_filter: Optional callsign substring filter\n    start_date: Optional start date filter\n    end_date: Optional end date filter\n    db: Database session\n\nReturns:\n    Paginated list of QSOs",
        "operationId": "list_qsos_api_v1_qso_list_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "callsign_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Callsign Filter"
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QSOListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Qsos",
        "tags": [
          "QSO Logging",
          "qso"
        ]
      }
    },
    "/api/v1/qso/log": {
      "post": {
        "description": "Log a QSO with auto-populated fields from image metadata.\n\nAuto-populates fields from image metadata using fallback hierarchy:\n1. User override (from request)\n2. Image metadata (from decode)\n3. None/empty\n\nArgs:\n    request: QSO logging request with image_id and optional overrides\n    db: Database session\n\nReturns:\n    Created QSO record\n\nRaises:\n    HTTPException: If image not found or callsign missing/invalid",
        "operationId": "log_qso_api_v1_qso_log_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogQSORequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QSOResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log Qso",
        "tags": [
          "QSO Logging",
          "qso"
        ]
      }
    },
    "/api/v1/qso/{qso_id}": {
      "delete": {
        "description": "Delete QSO by ID.\n\nArgs:\n    qso_id: QSO ID\n    db: Database session\n\nRaises:\n    HTTPException: If QSO not found",
        "operationId": "delete_qso_api_v1_qso__qso_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "qso_id",
            "required": true,
            "schema": {
              "title": "Qso Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Qso",
        "tags": [
          "QSO Logging",
          "qso"
        ]
      },
      "get": {
        "description": "Get single QSO by ID.\n\nArgs:\n    qso_id: QSO ID\n    db: Database session\n\nReturns:\n    QSO record\n\nRaises:\n    HTTPException: If QSO not found",
        "operationId": "get_qso_api_v1_qso__qso_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "qso_id",
            "required": true,
            "schema": {
              "title": "Qso Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QSOResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Qso",
        "tags": [
          "QSO Logging",
          "qso"
        ]
      }
    },
...
  }
}